Create Divisible Sum Pairs
You are given an array of  integers, , and a positive integer, . Find and print the number of  pairs where  and  +  is divisible by . 

For example,  and . Our three pairs meeting the criteria are  and . 

Function Description

Complete the divisibleSumPairs function in the editor below. It should return the integer count of pairs meeting the criteria. 

divisibleSumPairs has the following parameter(s): 
•n: the integer length of array  
•ar: an array of integers 
•k: the integer to divide the pair sum by 



Input Format



The first line contains  space-separated integers,  and . 
 The second line contains  space-separated integers describing the values of