Strong Password Solution in Python
Louise joined a social networking site to stay in touch with her friends. The signup page required her to input a name and a password. However, the password must be strong. The website considers a password to be strong if it satisfies the following criteria:
•Its length is at least . 
•It contains at least one digit. 
•It contains at least one lowercase English character. 
•It contains at least one uppercase English character. 
•It contains at least one special character. The special characters are: !@#$%^&*()-+ 

She typed a random string of length  in the password field but wasn't sure if it was strong. Given the string she typed, can you find the minimum number of characters she must add to make her password strong?

Note: Here's the set of types of characters in a form you can paste in your solution:

numbers = "0123456789"
lower_case = "abcdefghijklmnopqrstuvwxyz"
upper_case = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
special_characters = "!@#$%^&*()-+"




Input Format



The first line contains an integer  denoting the length of the string.

The second line contains a string consisting of  characters, the password typed by Louise. Each character is either a lowercase/uppercase English alphabet, a digit, or a special character.



Constraints


 • 



Output Format



Print a single line containing a single integer denoting the answer to the problem.



Sample Input 0



3
Ab1




Sample Output 0



3